+++ src/StateSmith/Runner/DiServiceProvider.cs
@@ -170,6 +170,7 @@ public class DiServiceProvider : IDisposable
         services.AddSingleton(renderConfigAllVars.C);
         services.AddSingleton(renderConfigAllVars.CSharp);
         services.AddSingleton(renderConfigAllVars.JavaScript);
+        services.AddSingleton(renderConfigAllVars.Java);
         services.AddSingleton<IExpansionVarsPathProvider, CSharpExpansionVarsPathProvider>();
         services.AddSingleton<RunnerSettings>(new RunnerSettings(""));
         services.AddSingleton<FilePathPrinter>(new FilePathPrinter(""));